+++ PopupBridge/src/main/java/com/braintreepayments/popupbridge/PopupBridge.java
@@ -98,7 +98,15 @@ public class PopupBridge extends BrowserSwitchFragment {
         String error = null;
         String payload = null;
 
-        if (result == BrowserSwitchResult.OK) {
+        if (result == BrowserSwitchResult.CANCELED) {
+            mWebView.evaluateJavascript(""
+                + "if (typeof window.popupBridge.onCancel === 'function') {"
+                + "  window.popupBridge.onCancel();"
+                + "} else {"
+                + "  window.popupBridge.onComplete(null, null);"
+                + "}", null);
+            return;
+        } else if (result == BrowserSwitchResult.OK) {
             if (returnUri == null || !returnUri.getScheme().equals(getReturnUrlScheme()) ||
                     !returnUri.getHost().equals(POPUP_BRIDGE_URL_HOST)) {
                 return;
@@ -125,14 +133,6 @@ public class PopupBridge extends BrowserSwitchFragment {
             } catch (JSONException ignored) {}
 
             payload = json.toString();
-        } else if (result == BrowserSwitchResult.CANCELED) {
-            mWebView.evaluateJavascript(""
-                + "if (typeof window.popupBridge.onClose === 'function') {"
-                + "  window.popupBridge.onClose();"
-                + "} else {"
-                + "  window.popupBridge.onComplete(null, null);"
-                + "}", null);
-            return;
         } else if (result == BrowserSwitchResult.ERROR) {
             error = "new Error('" + result.getErrorMessage() + "')";
         }