+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## unreleased
 
-* If page has created an `popupBridge.onClose` function, it will be called when user closes the window
+* If page has created an `popupBridge.onCancel` function, it will be called when user closes the window
 
 ## 0.1.1 (2017-03-30)